+++ test/case/atbus_node_setup_test.cpp
@@ -38,7 +38,8 @@ static int node_setup_test_on_error(const atbus::node &n, const atbus::endpoint 
   std::streamsize w = std::cout.width();
   CASE_MSG_INFO() << "[Log Error] node=0x" << std::setfill('0') << std::hex << std::setw(8) << n.get_id() << ", ep=0x"
                   << std::setw(8) << (nullptr == ep ? 0 : ep->get_id()) << ", c=" << conn << std::setfill(' ')
-                  << std::setw(w) << std::dec << "=> status: " << status << ", errcode: " << errcode << std::endl;
+                  << std::setw(static_cast<int>(w)) << std::dec << "=> status: " << status << ", errcode: " << errcode
+                  << std::endl;
   return 0;
 }
 
@@ -47,7 +48,8 @@ static int node_setup_test_on_info_log(const atbus::node &n, const atbus::endpoi
   std::streamsize w = std::cout.width();
   CASE_MSG_INFO() << "[Log Info] node=0x" << std::setfill('0') << std::hex << std::setw(8) << n.get_id() << ", ep=0x"
                   << std::setw(8) << (nullptr == ep ? 0 : ep->get_id()) << ", c=" << conn << std::setfill(' ')
-                  << std::setw(w) << std::dec << "=> message: " << (nullptr == msg ? "" : msg) << std::endl;
+                  << std::setw(static_cast<int>(w)) << std::dec << "=> message: " << (nullptr == msg ? "" : msg)
+                  << std::endl;
 
   return 0;
 }
